fix: cambia la forma en la que se calcula el TotalLabor
No se porque estaba casteando el costo de la hora de trabajo a float

+++ ShopMGR.Aplicacion/Servicios/AdministracionTrabajos.cs
@@ -159,7 +159,7 @@ namespace ShopMGR.Aplicacion.Servicios
             if (!trabajo.TotalLabor.HasValue)
             {
                 var costoHora = await _repositorioPresupuestos.ObtenerCostoHoraDeTrabajo();
-                trabajo.TotalLabor = (decimal)(trabajo.HorasDeTrabajo.Sum(h => h.Horas) * (float)costoHora);
+                trabajo.TotalLabor = (decimal)trabajo.HorasDeTrabajo.Sum(h => h.Horas) * costoHora;
             }
 
             trabajo.FechaFin = DateOnly.FromDateTime(DateTime.Now);
